Update README.md
This project represents a significant journey of research and study, aimed at creating an insightful visualization of the Census Regions and Divisions of all the states in the USA. Leveraging Power BI, I have developed an interactive analysis that showcases the geographical and demographic distribution across the country. This project not only deepened my understanding of Census data but also enhanced my skills in data visualization and analysis.